+++ lib/index.d.ts
@@ -264,9 +264,11 @@ declare namespace Realm.Sync {
         readonly server: string;
         readonly token: string;
         static adminUser(adminToken: string, server?: string): User;
+        static login(server: string, username: string, password: string): Promise<Realm.Sync.User>;
         static login(server: string, username: string, password: string, callback: (error: any, user: User) => void): void;
-        static loginWithProvider(server: string, provider: string, providerToken: string, callback: (error: any, user: User) => void): void;
+        static register(server: string, username: string, password: string): Promise<Realm.Sync.User>;
         static register(server: string, username: string, password: string, callback: (error: any, user: User) => void): void;
+        static registerWithProvider(server: string, options: { provider: string, providerToken: string, userInfo: any }): Promise<Realm.Sync.User>;
         static registerWithProvider(server: string, options: { provider: string, providerToken: string, userInfo: any }, callback: (error: Error | null, user: User | null) => void): void;
         logout(): void;
         openManagementRealm(): Realm;
@@ -375,7 +377,9 @@ declare class Realm {
      * @param {Configuration} config 
      */
     static open(config: Realm.Configuration): Promise<Realm>
+
     /**
+     * @deprecated in favor of `Realm.open`
      * Open a realm asynchronously with a callback. If the realm is synced, it will be fully synchronized before it is available.
      * @param {Configuration} config 
      * @param {Function} callback will be called when the realm is ready.