+++ tests/js/session-tests.js
@@ -50,32 +50,6 @@ function uuid() {
     });
 }
 
-function promisifiedRegister(server, username, password) {
-    return new Promise((resolve, reject) => {
-        Realm.Sync.User.register(server, username, password, (error, user) => {
-            if (error) {
-                console.log(`promisifiedRegister ${error}`);
-                reject(error);
-            } else {
-                resolve(user);
-            }
-        });
-    });
-}
-
-function promisifiedLogin(server, username, password) {
-    return new Promise((resolve, reject) => {
-        Realm.Sync.User.login(server, username, password, (error, user) => {
-            if (error) {
-                console.log(`promisifiedLogin ${error}`);
-                reject(error);
-            } else {
-                resolve(user);
-            }
-        });
-    });
-}
-
 function runOutOfProcess(nodeJsFilePath) {
     var nodeArgs = Array.prototype.slice.call(arguments);
     let tmpDir = tmp.dirSync();
@@ -101,7 +75,7 @@ module.exports = {
     },
 
     testProperties() {
-        return promisifiedRegister('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.login('http://localhost:9080', uuid(), 'password').then(user => {
             return new Promise((resolve, reject) => {
 
                 const accessTokenRefreshed = this;
@@ -152,7 +126,7 @@ module.exports = {
 
         runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
             .then(() => {
-                return promisifiedLogin('http://localhost:9080', username, 'password').then(user => {
+                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user => {
                     const accessTokenRefreshed = this;
                     let successCounter = 0;
 
@@ -188,7 +162,7 @@ module.exports = {
 
         runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
             .then(() => {
-                return promisifiedLogin('http://localhost:9080', username, 'password').then(user => {
+                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user => {
                     return new Promise((resolve, reject) => {
                         const accessTokenRefreshed = this;
                         let successCounter = 0;
@@ -241,7 +215,7 @@ module.exports = {
 
         runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
             .then(() => {
-                return promisifiedLogin('http://localhost:9080', username, 'password').then(user => {
+                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user => {
                     const accessTokenRefreshed = this;
                     let successCounter = 0;
                     let progressNotificationCalled = false;
@@ -277,7 +251,7 @@ module.exports = {
 
         runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
             .then(() => {
-                return promisifiedLogin('http://localhost:9080', username, 'password').then(user => {
+                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user => {
                     return new Promise((resolve, reject) => {
                         let progressNotificationCalled = false;
                         let config = {
@@ -324,7 +298,7 @@ module.exports = {
 
        runOutOfProcess(__dirname + '/download-api-helper.js', username, realmName, REALM_MODULE_PATH)
             .then(() => {
-                return promisifiedLogin('http://localhost:9080', username, 'password').then(user => {
+                return Realm.Sync.User.login('http://localhost:9080', username, 'password').then(user => {
                     return new Promise((resolve, reject) => {
                         const accessTokenRefreshed = this;
                         let successCounter = 0;
@@ -429,7 +403,7 @@ module.exports = {
     },
 
     testErrorHandling() {
-        return promisifiedRegister('http://localhost:9080', uuid(), 'password').then(user => {
+        return Realm.Sync.User.register('http://localhost:9080', uuid(), 'password').then(user => {
             return new Promise((resolve, _reject) => {
                 const config = { sync: { user, url: 'realm://localhost:9080/~/myrealm' } };
                 config.sync.error = (sender, error) => {