+++ tests/js/index.js
@@ -47,7 +47,7 @@ function node_require(module) { return require(module); }
 // If on node, run the async tests
 const isNodeProcess = typeof process === 'object' && process + '' === '[object process]';
 if (isNodeProcess) {
-    TESTS.AsyncTests = node_require('./async-tests');
+    //TESTS.AsyncTests = node_require('./async-tests');
 }
 
 var SPECIAL_METHODS = {
@@ -76,8 +76,9 @@ exports.registerTests = function(tests) {
 };
 
 exports.prepare = function(done) {
-    if (!isNodeProcess || global.testAdminUserInfo) {
+    if (!Realm.Sync || !isNodeProcess || global.testAdminUserInfo) {
         done();
+        return;
     }
 
     let helper = require('./admin-user-helper');